Use floats for position and velocity

+++ src/math_helper.rs
@@ -1,9 +1,9 @@
-use std::f64::consts;
+use std::f32::consts;
 
 use rand;
 
-pub fn get_point_on_unit_circle() -> (f64, f64)
+pub fn get_point_on_unit_circle() -> (f32, f32)
 {
-    let theta: f64 = rand::random::<f64>() * 2.0 * consts::PI;
+    let theta: f32 = rand::random::<f32>() * 2.0 * consts::PI;
     (theta.cos(), theta.sin())
 }
